refactor: rename actions.js -> utils.js
